+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 
 # k6 IntelliJ Plugin Changelog
 
-## [Unreleased]
+## [0.1.0]
+### Added
+- k6 Run configuration
+- create run configuration from file or function
+- show pass/failed run in the test console
+- console output with colors 
+### Fixed
+- default shortcut for the "run file locally" action was changed from 'shift+K' to 'ctrl+shift+K' so as not conflicting when typing a capital 'K' in the editor 
 ## [0.0.1]
 ### Added
 - Run script locally / in the cloud